pyproject.toml
[project]
name = "ai-photos-management"
version = "0.1.0"
description = "AI-powered photo management system using SOTA models"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "torch>=2.9.0",
    "torchvision",
    "transformers>=4.57.1",
    "pillow>=12.0.0",
    "accelerate>=1.11.0",
    "flask>=3.1.2",
    "sqlalchemy>=2.0.44",
    "psycopg2-binary>=2.9.11",
    "pgvector>=0.4.1",
    "celery>=5.5.3",
    "redis>=7.0.1",
    "open-clip-torch>=3.2.0",
    "paddleocr>=3.3.1",
    "insightface>=0.7.3",
    "opencv-python>=4.12.0.88",
    "pdqhash>=0.2.8",
    "scikit-learn>=1.7.2",
    "tqdm>=4.67.1",
    "python-dotenv>=1.2.1",
    "pillow-heif>=1.1.1",
    "rawpy>=0.25.1",
    "paddlepaddle-gpu>=3.2.1",
    "onnxruntime-gpu>=1.23.2",
]

[tool.uv.sources]
# Install PyTorch with CUDA support on Linux/Windows (CUDA doesn't exist for Mac).
# NOTE: We must explicitly request them as `dependencies` above. These improved
# versions will not be selected if they're only third-party dependencies.
torch = [
  { index = "pytorch-cuda", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchaudio = [
  { index = "pytorch-cuda", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cuda", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

paddlepaddle-gpu = [
  { index = "paddlepaddle-cuda", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "pytorch-cuda"
# Use PyTorch built for NVIDIA Toolkit version 13.0.
# Available versions: https://pytorch.org/get-started/locally/
url = "https://download.pytorch.org/whl/cu130"
# Only use this index when explicitly requested by `tool.uv.sources`.
explicit = false

# NOTE: No ARM64 build yet
[[tool.uv.index]]
name = "paddlepaddle-cuda"
# Use PaddlePaddle built for NVIDIA Toolkit version 13.0.
# Available versions: https://www.paddlepaddle.org.cn/install/quick?docurl=/documentation/docs/zh/develop/install/pip/linux-pip.html
url = "https://www.paddlepaddle.org.cn/packages/stable/cu130"
# Only use this index when explicitly requested by `tool.uv.sources`.
explicit = false
